Add class to rivet routine which divides hists by channel
Added a wrapper class for histogram production which produces both a
total histogram and histograms by channel simultaneously. Of course, not
all histograms will always be filled by this makes it more clear which
channels were included in the run. This can easily be extended by add
more channels or the option to use less by histogram.